Added basic card template prefab and scriptable object
Will add more later but basic template is made. It references it's elements and contains a slot for a scriptable object that contains all of the values you want into. We can just use some default settings and change small things, also the data in the template will be updated as we add more things like status effects. Also I added a sprite pack that has a few assets for making cards that I used