Added Energy UI and Fixed some UI issues
Added a small circle UI element in the bottom left that displays your energy for that turn and changes based on the energy value of the card or cards that you played. I also noticed some issues with the UI which made it format incorrectly and I have attempted to fix them.