Fixed options menu and scene management
Fixed the layering issue with the options so the buttons are now selectable, also fixed a bug where options menu had the incorrect load input and would load tutorial scene. Adjusted the close button to properly close and load the last scene. Also worked on the scene management so that it records the current and last scene properly, so far it's only used on the close options button however since we do not have a load game mechanic